Create saddle_point.cpp
This program finds the saddle point in a matrix. It first finds the minimum in each row and restored them in a vector. Then finds whether the minimum is also the maximum value in each column. The difficulties lies in realizing that there can exists multiple minimum or maximum in each row or column