Create find_mode_in array.cpp
Find mode in an array. It is implemented using the map template. There may be potential bugs when it is operated on an array that contains 0. Because when two elements have the same frequency, the max_key will be set to null and it will be converted to 0 automatically. When 0 is the mode, the program may fail to provide the correct mode.